chore(ffi): move replies to their own file
